Move 'Product list' component and styling to a separate folder.
