+++ client/my-sites/plans/jetpack-plans/product-store/product-list/index.tsx
@@ -1,15 +1,17 @@
 import { useTranslate } from 'i18n-calypso';
 import { useMemo } from 'react';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
-import productButtonLabel from '../product-card/product-button-label';
-import { FeaturedItemCard } from './featured-item-card';
-import { HeroImage } from './hero-image';
-import { useProductsToDisplay } from './hooks/use-products-to-display';
-import { useStoreItemInfo } from './hooks/use-store-item-info';
-import { MostPopular } from './most-popular';
-import SimpleProductCard from './simple-product-card';
-import { getSortedDisplayableProducts } from './utils/get-sorted-displayable-products';
-import type { ProductsListProps } from './types';
+import productButtonLabel from '../../product-card/product-button-label';
+import { FeaturedItemCard } from '../featured-item-card';
+import { HeroImage } from '../hero-image';
+import { useProductsToDisplay } from '../hooks/use-products-to-display';
+import { useStoreItemInfo } from '../hooks/use-store-item-info';
+import { MostPopular } from '../most-popular';
+import SimpleProductCard from '../simple-product-card';
+import { getSortedDisplayableProducts } from '../utils/get-sorted-displayable-products';
+import type { ProductsListProps } from '../types';
+
+import './style.scss';
 
 export const ProductsList: React.FC< ProductsListProps > = ( {
 	createCheckoutURL,
