+++ client/my-sites/plans/jetpack-plans/product-store/product-list/style.scss
@@ -1,0 +1,36 @@
+@import '@wordpress/base-styles/breakpoints';
+@import '@wordpress/base-styles/mixins';
+
+.jetpack-product-store__products-list-all {
+	margin-top: 4em;
+	margin-bottom: 48px;
+}
+
+.jetpack-product-store__products-list-all h3 {
+	margin-bottom: 2.5em;
+}
+
+.jetpack-product-store__products-list-all-header {
+	margin-top: rem( 32px );
+	margin-bottom: rem( 24px );
+	font-size: $font-title-small;
+	font-weight: 700;
+
+	@include break-medium {
+		margin-top: rem( 72px );
+		margin-bottom: rem( 48px );
+		font-size: $font-title-medium;
+		line-height: rem( 29px );
+		font-weight: 600;
+	}
+}
+
+.jetpack-product-store__products-list-all-grid {
+	display: grid;
+	column-gap: 80px;
+	row-gap: 32px;
+
+	@include break-medium {
+		grid-template-columns: repeat( 2, 1fr );
+	}
+}
