+++ client/my-sites/plans/jetpack-plans/product-store/style.scss
@@ -83,46 +83,12 @@
 		margin-inline-end: 4px;
 	}
 
-	&__products-list-all {
-		margin-top: 4em;
-		margin-bottom: 48px;
-	}
-
-	&__products-list-all h3 {
-		margin-bottom: 2.5em;
-	}
-
 	.button {
 		color: var( --color-neutral-100 );
 		border: 1px solid var( --color-neutral-100 );
 		font-family: 'SF Pro Text', $sans;
 		font-size: $font-body;
 		font-weight: 600;
-	}
-}
-
-.jetpack-product-store__products-list-all-header {
-	margin-top: rem( 32px );
-	margin-bottom: rem( 24px );
-	font-size: $font-title-small;
-	font-weight: 700;
-
-	@include break-medium {
-		margin-top: rem( 72px );
-		margin-bottom: rem( 48px );
-		font-size: $font-title-medium;
-		line-height: rem( 29px );
-		font-weight: 600;
-	}
-}
-
-.jetpack-product-store__products-list-all-grid {
-	display: grid;
-	column-gap: 80px;
-	row-gap: 32px;
-
-	@include break-medium {
-		grid-template-columns: repeat( 2, 1fr );
 	}
 }
